fix: added ColorPicker component
