+++ packages/_storybook_/package.json
@@ -18,6 +18,7 @@
     "@reciprocity/day-selector": "^0.1.4",
     "@reciprocity/loading-spinner": "^0.1.5",
     "@reciprocity/multiselect": "^0.2.2",
+    "@reciprocity/color-picker": "^0.1.0",
     "bootstrap": "^4.3.1",
     "bootstrap-vue": "^2.0.2",
     "vue": "^2.6.10"
