+++ packages/color-picker/index.js
@@ -1,0 +1,2 @@
+import ColorPicker from "./ColorPicker.vue";
+export default ColorPicker;
