+++ packages/color-picker/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@reciprocity/color-input",
+  "name": "@reciprocity/color-picker",
   "version": "0.1.6",
-  "description": "An input field for color values that includes a color picker",
+  "description": "A color picker where you can select the color or type the relevant HEX value.",
   "main": "dist/index.js",
   "scripts": {},
   "author": "",
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "@reciprocity/bootstrap": "^0.6.0",
-    "@reciprocity/color-picker": "^0.1.0",
     "@simonwep/pickr": "^1.2.7"
   }
 }
